+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Minor UI improvements (updated dark primary btn contrast).
 
+- Writes the status header for JSON responses only if the fields picker succeed or has acceptable fallback.
+    _This is to allow custom response status code in reaction to failed json writes._
+
 - Bumped `golang.org/x/*` dependencies to silence security scanners ([#7829](https://github.com/pocketbase/pocketbase/discussions/7829)).